+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Additional License Descriptors
+- Added JSON Schema for the licenses export available
+  on [https://www.cmgapps.com/schema/licenses-schema.json](https://www.cmgapps.com/schema/licenses-schema.json)
 
 ### Changed
 
@@ -109,7 +111,7 @@ Use Gradle's Reports API to generate reports
 
 ### Changed
 
-- 
+-
     - `<version>` is now attribute to `<library>`
     - `<url>` is now a attribute to `<license>`
     - `<library>` has an `id` attribute now
@@ -160,18 +162,33 @@ Use Gradle's Reports API to generate reports
 - Use Kotlin Serialization instead of Moshi
 
 [Unreleased]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/5.1.0...HEAD
+
 [5.1.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/5.0.0...5.1.0
+
 [5.0.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.8.0...5.0.0
+
 [4.8.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.7.0...4.8.0
+
 [4.7.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.6.1...4.7.0
+
 [4.6.1]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.6.0...4.6.1
+
 [4.6.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.5.0...4.6.0
+
 [4.5.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.4.0...4.5.0
+
 [4.4.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.3.0...4.4.0
+
 [4.3.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.2.0...4.3.0
+
 [4.2.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.1.0...4.2.0
+
 [4.1.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/4.0.0...4.1.0
+
 [4.0.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/3.3.0...4.0.0
+
 [3.3.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/3.2.0...3.3.0
+
 [3.2.0]: https://github.com/chrimaeon/gradle-licenses-plugin/compare/3.1.0...3.2.0
+
 [3.1.0]: https://github.com/chrimaeon/gradle-licenses-plugin/commits/3.1.0